html/semantics/sections/the-body-element/sections-body-element.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>Body Element</title>
    <link rel="author" title="dzenana" href="mailto:dzenana.trenutak@gmail.com">
    <link rel="help" href="http://www.w3.org/html/wg/drafts/html/CR/single-page.html#the-body-element">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body id="body_here">

    <div id="log"></div>

    <script>
        "use strict";
        // check that the document.body IDL attribute provides scripts with access to the body element
        test(function () {
            assert_equals(document.body, document.getElementById("body_here"), "document.body accesses the body element");
        }, "document.body accesses the body element");
    </script>
	
</body>
</html>
